🧬 Rename the sub-package about subcommand line *get* to be more clear its meaning.
