+++ test/unit_test/command/get/component.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from pymock_api.command.inspect.component import SubCmdGetComponent
+from pymock_api.command.get.component import SubCmdGetComponent
 from pymock_api.model.cmd_args import SubcmdGetArguments
